requirements.txt
Django==1.6.1
django-tastypie==0.11.0
django-pipeline==1.3.22
requests==2.2.1
